refactor: separate comment according to FSD features layer
