--- apps/web/src/features/blog/comment/api/actions/deleteCommentAction.ts
@@ -1,10 +1,0 @@
-'use server';
-
-import { trpc } from '@/fsd/shared/api/trpc/server';
-import { revalidatePath } from 'next/cache';
-
-export async function deleteCommentAction(commentId: string, postId: string) {
-	await trpc.comment.deleteComment(commentId);
-
-	revalidatePath(`/blog/${postId}`);
-}
