--- apps/web/src/features/blog/comment/api/actions/toggleLikeCommentAction.ts
@@ -1,18 +1,0 @@
-'use server';
-
-import { trpc } from '@/fsd/shared/api/trpc/server';
-import { revalidatePath } from 'next/cache';
-
-export async function toggleLikeCommentAction(
-	commentId: string,
-	postId: string,
-	userId: string,
-) {
-	const updatedComment = await trpc.comment.toggleLike({
-		commentId,
-		userId: userId,
-	});
-
-	revalidatePath(`/blog/${postId}`);
-	return updatedComment;
-}
