--- apps/web/src/features/blog/comment/api/actions/updateCommentAction.ts
@@ -1,18 +1,0 @@
-'use server';
-
-import { trpc } from '@/fsd/shared/api/trpc/server';
-import { revalidatePath } from 'next/cache';
-
-export async function updateCommentAction(
-	commentId: string,
-	content: string,
-	postId: string,
-) {
-	const updatedComment = await trpc.comment.updateComment({
-		id: commentId,
-		content,
-	});
-
-	revalidatePath(`/blog/${postId}`);
-	return updatedComment;
-}
