--- apps/web/src/features/blog/comment/api/index.ts
@@ -1,5 +1,0 @@
-export { useGetComments } from './queries/useGetComments';
-export { createCommentAction } from './actions/createCommentAction';
-export { toggleLikeCommentAction } from './actions/toggleLikeCommentAction';
-export { deleteCommentAction } from './actions/deleteCommentAction';
-export { updateCommentAction } from './actions/updateCommentAction';
