--- apps/web/src/features/blog/comment/index.ts
@@ -1,4 +1,0 @@
-export * from './ui';
-export * from './api';
-export * from './model';
-export * from './lib';
