--- apps/web/src/features/blog/comment/lib/calculateCommentCount.ts
@@ -1,7 +1,0 @@
-import type { Comment } from '@jung/shared/types';
-
-export const calculateCommentCount = (comments: Comment[]) =>
-	comments.reduce(
-		(count, comment) => count + 1 + (comment.replies?.length || 0),
-		0,
-	);
