--- apps/web/src/features/blog/comment/lib/index.ts
@@ -1,3 +1,0 @@
-export { calculateCommentCount } from './calculateCommentCount';
-export * from './commentOptimisticUpdates';
-export { transformComments } from './transformComments';
