--- apps/web/src/features/blog/comment/model/index.ts
@@ -1,5 +1,0 @@
-export { useCreateComment } from './useCreateComment';
-export { useDeleteComment } from './useDeleteComment';
-export { useUpdateComment } from './useUpdateComment';
-export { useToggleLikeComment } from './useToggleLikeComment';
-export { useInfiniteScroll } from './useInfiniteScroll';
