--- apps/web/src/features/blog/comment/model/useInfiniteScroll.ts
@@ -1,19 +1,0 @@
-'use client';
-
-import { useEffect } from 'react';
-import { useInView } from 'react-intersection-observer';
-
-export const useInfiniteScroll = (
-	hasNextPage: boolean,
-	fetchNextPage: () => void,
-) => {
-	const { ref, inView } = useInView();
-
-	useEffect(() => {
-		if (inView && hasNextPage) {
-			fetchNextPage();
-		}
-	}, [inView, fetchNextPage, hasNextPage]);
-
-	return ref;
-};
