--- apps/web/src/features/blog/comment/ui/index.ts
@@ -1,4 +1,0 @@
-export { default as CommentItem } from './CommentItem';
-export { default as CommentList } from './CommentList';
-export { default as CommentForm } from './CommentForm';
-export * from './CommentList.css';
