+++ apps/web/src/features/comment/index.ts
@@ -1,0 +1,1 @@
+export { ViewComments } from './viewComments/ui/ViewComments';
